refactor: inline Meetuser interface into payload types and remove unused user types file
