+++ src/tencent-mtg-hook/types/index.ts
@@ -30,5 +30,3 @@ export * from './speaker.types';
 // 录制数据类型定义
 export * from './recording-data.types';
 
-// 用户相关类型定义
-export * from './user.types';
